+++ CHANGES.rst
@@ -19,6 +19,10 @@ Internal changes
 * ``obs_suie``: name log files according to the date of the source files instead of simply numbering them consecutively (:issue:`74`, :pull:`79`)
 * ``obs_suite``: standardize level scripts (:pull:`79`)
 
+Breaking changes
+^^^^^^^^^^^^^^^^
+* ``obs_suite``: rename level3 output from <YYYY>-<MM>-<RELEASE>-<UPDATE>-pressure_data.psv to pressure-data-<YYYY>-<MM>-<RELEASE>-<UPDATE>.psv (:pull:`79`)
+
 v7.1.0 (2024-11-25)
 -------------------
 Contributors to this version: Ludwig Lierhammer (:user:`ludwiglierhammer`)